The Duck and the Fox
A duck is in the center of a circular pond. At the edge of the pond
there is a Fox that wants to eat the duck. The duck cannot take off from
the water, it has to get to land to be able to fly away. The Fox can't
swim, but can run four times as fast as the Duck swims. Can the Duck get
away?